sensorMaster/requirements.txt
pyserial == 3.1.1
pyzmq == 16.0.0
